requirements.txt
# Minimal deps for local helper scripts in ./scripts
PyYAML>=6.0
